requirements.txt
click==8.1.7
PyGithub==2.5.0
pytest==8.3.4
Requests==2.32.3
